HPCC-34154: Move index event summary to shared library
